+++ internal/runtime/docker/container_start.go
@@ -12,7 +12,7 @@ import (
 	"github.com/docker/go-connections/nat"
 )
 
-func (docker *Docker) Start(ctx context.Context, options runtimeTypes.ContainerOptions) error {
+func (docker *Docker) Up(ctx context.Context, options runtimeTypes.ContainerOptions, code bool) error {
 
 	if options.ImageName == "" {
 		options.ImageName = "envcontainer/" + options.ContainerName
@@ -49,13 +49,18 @@ func (docker *Docker) Start(ctx context.Context, options runtimeTypes.ContainerO
 		}
 
 		options.Commands = []string{getContainer.Command}
+
+		if code {
+			return docker.code(ctx, getContainer.ID, options)
+		}
+
 		return docker.exec(ctx, getContainer.ID, options)
 	}
 
-	return docker.tryCreateAndStartContainer(ctx, options)
+	return docker.tryCreateAndStartContainer(ctx, options, code)
 }
 
-func (docker *Docker) containerCreateAndStart(ctx context.Context, options runtimeTypes.ContainerOptions) error {
+func (docker *Docker) containerCreateAndStart(ctx context.Context, options runtimeTypes.ContainerOptions, code bool) error {
 
 	exposedPorts := nat.PortSet{}
 	portBindings := nat.PortMap{}
@@ -82,25 +87,22 @@ func (docker *Docker) containerCreateAndStart(ctx context.Context, options runti
 		}
 	}
 
-	// mounts := options.Mounts
-	// mounts = append(mounts)
-
 	mounts := docker.buildMount(options.DefaultMountDir, options.Mounts)
 
 	bindProject := options.HostDirToBind + ":/home/" + options.ContainerName
 
 	// Create the container
 	containerResponse, err := docker.client.ContainerCreate(ctx, &container.Config{
-		User:         options.User,
 		WorkingDir:   "/home/" + options.ContainerName,
 		Image:        options.ImageName,
 		ExposedPorts: exposedPorts,
 		Tty:          true,
-		Cmd:          options.Commands,
+		Hostname:     options.ContainerName,
 	}, &container.HostConfig{
 		PortBindings: portBindings,
 		Binds:        []string{bindProject},
 		Mounts:       mounts,
+		NetworkMode:  network.NetworkBridge,
 	}, &network.NetworkingConfig{}, nil, options.ContainerName)
 	if err != nil {
 		return err
@@ -112,6 +114,10 @@ func (docker *Docker) containerCreateAndStart(ctx context.Context, options runti
 		ID:   containerResponse.ID,
 	}, container.StartOptions{})
 
+	if code {
+		return docker.code(ctx, containerResponse.ID, options)
+	}
+
 	return docker.exec(ctx, containerResponse.ID, options)
 }
 
@@ -119,23 +125,24 @@ func (docker *Docker) tryStart(ctx context.Context, info runtimeTypes.ContainerS
 	err := docker.client.ContainerStart(ctx, info.ID, options)
 	if err != nil {
 		fmt.Print("Error to start container, ")
-		docker.Stop(ctx, runtimeTypes.ContainerOptions{
+		docker.Down(ctx, runtimeTypes.ContainerOptions{
 			ContainerName: info.Name,
 		})
 		return err
 	}
+
 	return nil
 }
 
-func (docker *Docker) tryCreateAndStartContainer(ctx context.Context, options runtimeTypes.ContainerOptions) error {
+func (docker *Docker) tryCreateAndStartContainer(ctx context.Context, options runtimeTypes.ContainerOptions, code bool) error {
 
 	if len(options.Commands) > 0 && options.Commands[0] != "" {
-		return docker.containerCreateAndStart(ctx, options)
+		return docker.containerCreateAndStart(ctx, options, code)
 	}
 	var err error
 	for _, shell := range shells {
 		options.Commands = []string{shell}
-		if err = docker.containerCreateAndStart(ctx, options); err == nil {
+		if err = docker.containerCreateAndStart(ctx, options, code); err == nil {
 			return nil
 		}
 	}