+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ### Added
 
 - chore: remove volume when stop container
+- chore: remove option user on .envcontainer.yaml
+- chore: add new flag to start with vscode
+- fix: regex to validate ports pattern
 
 ### Fixed